Move viewModel to separate file
